+++ satrap/frontend/__init__.py
@@ -3,5 +3,5 @@
 __project__ = "satrap/frontend"
 
 CLI = "SATRAP"
-VERSION = f"{__project__ } v0.1"
+VERSION = f"{__project__ } v0.1.1"
 DESCRIPTION = "SATRAP. Cyber threat Intelligence powered by automated reasoning."