+++ satrap/service/__init__.py
@@ -1,3 +1,6 @@
 """SATRAP CTI analysis package"""
 
 __project__ = "satrap/service"
+
+# Perhaps it'd be good to import these here for easier access in the service module.
+# from satrap.commons.format_utils import format_dict, tabulate_stix_obj, tabulate_groups